Added compose-override merging on install inside docker

+++ .internal/templates/scripts/build_installer.js
@@ -131,6 +131,9 @@ if [[ ! -f ./menu.sh ]]; then
   exit 2
 fi
 
+# Load meta data from installation
+source ./.internal/meta.sh
+
 #### Prebuild service scripts
 ${renderPrebuildScripts(options?.prebuildScripts)}
 
@@ -145,7 +148,14 @@ ${renderPostbuildScripts(options?.postbuildScripts)}
 
 #### End postbuild service scripts
 
-cp docker-compose-base.yml docker-compose.yml
+if [[ -f ./compose-override.yml ]]; then
+  echo "Merging 'compose-override.yml' with 'docker-compose-base.yml':"
+  echo "docker container run -it -v $(pwd)/compose-override.yml:/usr/iotstack_pycli/compose-override.yml:ro -v $(pwd)/docker-compose-base.yml:/usr/iotstack_pycli/docker-compose-base.yml:ro -v $(pwd)/docker-compose.yml:/usr/iotstack_pycli/docker-compose.yml -e \\"PYCLI_OVERRIDE_YML=compose-override.yml\\" -e \\"PYCLI_BASE_YML=docker-compose-base.yml\\" -e \\"PYCLI_OUTPUT_YML=docker-compose.yml\\" iostack_pycli:$VERSION /usr/local/bin/python3 /usr/iotstack_pycli/compose_override_entry.py"
+  docker container run -it -v $(pwd)/compose-override.yml:/usr/iotstack_pycli/compose-override.yml:ro -v $(pwd)/docker-compose-base.yml:/usr/iotstack_pycli/docker-compose-base.yml:ro -v $(pwd)/docker-compose.yml:/usr/iotstack_pycli/docker-compose.yml -e "PYCLI_OVERRIDE_YML=compose-override.yml" -e "PYCLI_BASE_YML=docker-compose-base.yml" -e "PYCLI_OUTPUT_YML=docker-compose.yml" iostack_pycli:$VERSION /usr/local/bin/python3 /usr/iotstack_pycli/compose_override_entry.py
+else
+  rm docker-compose.yml
+  cp docker-compose-base.yml docker-compose.yml
+fi
 
 if [[ -f ./postbuild.sh ]]; then
   echo "Running postbuild script (${(options?.selectedServices ?? []).join(', ')}):"
